+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.1-dev (changes since 0.2.1 release go here)
 
+### Features
+* Added `__str__` method to `skbio.stats.ordination.OrdinationResults`.
+
 ## Version 0.2.1 (2014-10-27)
 
 This is an alpha release of scikit-bio. At this stage, major backwards-incompatible API changes can and will happen. Unified I/O with the scikit-bio I/O registry was the focus of this release.